requirements.txt
langchain
langchain-google-genai
langgraph
faiss-cpu
python-dotenv
tiktoken
langchain-community